drone/requirements.txt
Cython==0.29.24
matplotlib==3.5.0
numpy==1.19.5
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pycocotools==2.0.3
pycuda==2020.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytools==2020.4.3
pytz==2020.4
scipy==1.7.3
six==1.15.0
tidecv==1.0.1
tqdm==4.46.1
typing_extensions==4.0.0
torchvision==0.10.0
torch==1.9.0
